+++ src/hotspot/share/services/nmtUsage.hpp
